Follow C spec by not using default argument promotion for variadic functions.

+++ common/infrastructure/MUByteSet.h
@@ -10,14 +10,14 @@
 }
 
 + (instancetype) byteSet;
-+ (instancetype) byteSetWithBytes: (const uint8_t) firstByte, ...;
++ (instancetype) byteSetWithBytes: (const unsigned int) firstByte, ...;
 + (instancetype) byteSetWithBytes: (const uint8_t * const) bytes length: (size_t) length;
 
 - (instancetype) initWithBytes: (const uint8_t * const) bytes length: (size_t) length;
 - (instancetype) initWithFirstByte: (const uint8_t) first remainingBytes: (va_list) bytes;
 
 - (void) addByte: (const uint8_t) byte;
-- (void) addBytes: (const uint8_t) firstByte, ...;
+- (void) addBytes: (const unsigned int) firstByte, ...;
 - (void) addFirstByte: (const uint8_t) firstByte remainingBytes: (va_list) bytes;
 - (BOOL) containsByte: (const uint8_t) byte;
 - (NSData *) dataValue;
